# reto-tecnico/go-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API que procesa y rota matrices rectangulares protegida con JWT.",
        "title": "API en Go - Interseguro Challenge",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/process-matrix": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Recibe una matriz de números, la rota 90 grados y la reenvía a Node.js",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Recibe y rota una matriz (Protegido con JWT)",
                "parameters": [
                    {
                        "description": "Matriz de entrada",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.MatrixRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Autentica al usuario y devuelve un token JWT válido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Iniciar sesión para obtener Token JWT",
                "parameters": [
                    {
                        "description": "Credenciales de acceso",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Credentials": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "main.MatrixRequest": {
            "type": "object",
            "properties": {
                "matrix": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribe 'Bearer ' seguido de tu token JWT generador en /login (Ejemplo: Bearer eyJhbGci...).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}